feat: fully implemented a basic example for CSP
