+++ utils/contentSecurityPolicy/index.ts
@@ -1,6 +1,4 @@
-export { default as presetNextJS } from "./presetNextJS";
-export { default as presetNextJSPreview } from "./presetNextJSPreview";
 export { default as presetVercelToolbar } from "./presetVercelToolbar";
 export { default as presetVercelSpeedInsights } from "./presetVercelSpeedInsights";
 export { default as presetGoogleTagManager } from "./presetGoogleTagManager";
-export { default as presetGoogleTagManagerPreview } from "./presetGoogleTagManagerPreview";
+export { default as presetGoogleAnalytics } from "./presetGoogleAnalytics";
