+++ utils/contentSecurityPolicy/presetGoogleTagManager.ts
@@ -1,8 +1,8 @@
 import { CSPDirectives, nonce } from "csp-header";
 
-const preset: (hash: string) => Partial<CSPDirectives> = (hash) => ({
-    "script-src": ["https://www.googletagmanager.com", nonce(hash)],
+const preset: Partial<CSPDirectives> = {
+    "script-src": ["https://www.googletagmanager.com"],
     "connect-src": ["www.googletagmanager.com"],
     "img-src": ["www.googletagmanager.com"],
-});
+};
 export default preset;
