+++ utils/contentSecurityPolicy/presetVercelToolbar.ts
@@ -1,11 +1,11 @@
-import { BLOB, CSPDirectives, DATA, nonce, UNSAFE_INLINE } from "csp-header";
+import { BLOB, CSPDirectives, DATA, UNSAFE_INLINE } from "csp-header";
 
-const preset: (hash: string) => Partial<CSPDirectives> = (hash) => ({
-    "script-src": ["https://vercel.live", nonce(hash)],
+const preset: Partial<CSPDirectives> = {
+    "script-src": ["https://vercel.live"],
     "connect-src": ["https://vercel.live", "wss://ws-us3.pusher.com"],
     "img-src": ["https://vercel.live", "https://vercel.com", DATA, BLOB],
     "frame-src": ["https://vercel.live"],
-    "style-src": ["https://vercel.live", UNSAFE_INLINE /* , nonce(hash) */],
+    "style-src": ["https://vercel.live", UNSAFE_INLINE],
     "font-src": ["https://vercel.live", "https://assets.vercel.com"],
-});
+};
 export default preset;
